+++ spatial_index/builder.py
@@ -2,6 +2,8 @@ from . import _spatial_index as core
 from .index import SphereIndex
 from .io import write_sonata_meta_data_section
 
+import numpy as np
+
 
 class _WriteSONATAMetadataMixin:
     def _write_extended_meta_data_section(*a, **kw):
@@ -18,29 +20,19 @@ class _WriteSONATAMetadataMultiMixin:
         MPI.COMM_WORLD.Barrier()
 
 
-class SphereIndexBuilderBase:
-    def __init__(self):
-        self._core_builder = core.SphereIndex()
-
-    def add_sphere(self, id, center, radius):
-        self._core_builder._insert(id, center, radius)
-
+class SphereIndexBuilder:
     @classmethod
-    def create(cls, centroids, radii, output_dir=None):
+    def create(cls, centroids, radii, ids=None, output_dir=None):
         builder = cls()
 
-        for k, sphere in enumerate(zip(centroids, radii)):
-            builder.add_sphere(k, *sphere)
+        if ids is None:
+            ids = np.arange(centroids.shape[0])
+
+        builder._core_index = core.SphereIndex(centroids, radii, ids)
 
         builder._write_index_if_needed(output_dir)
         return builder._index_if_loaded
 
-    @classmethod
-    def from_numpy(cls, centroids, radii, output_dir=None):
-        return cls.create(centroids, radii)
-
-
-class SphereIndexBuilder(SphereIndexBuilderBase):
     def _write_index_if_needed(self, output_dir):
         self.index.write(output_dir)
 
@@ -50,4 +42,14 @@ class SphereIndexBuilder(SphereIndexBuilderBase):
 
     @property
     def index(self):
-        return SphereIndex(self._core_builder)
+        return SphereIndex(self._core_index)
+
+    @classmethod
+    def from_numpy(cls, centroids, radii, ids=None, output_dir=None):
+        return cls.create(centroids, radii, ids=ids, output_dir=output_dir)
+
+    def add_sphere(self, *a, **kw):
+        import warnings
+
+        warnings.warn("Deprecated in favour of 'from_numpy'.", DeprecationWarning)
+        raise RuntimeError("Invalid method, see deprecation warning.")