+++ tests/test_rtree_point.py
@@ -1,58 +1,49 @@
+import pytest
+
 import numpy as np
 
 from spatial_index import core
-
-from test_rtree_sphere import _box_query_id
-
-points = np.array(
-    [
-        [0.0, 1.0, 0],
-        [-0.5, -0.5, 0],
-        [0.5, -0.5, 0],
-        [-2.1, 0.0, 0],
-        [0.0, 2.1, 0],
-        [0.0, 0.0, 2.1],
-        [1., 1., 1.]
-    ], dtype=np.float32)
-
-ids = np.arange(len(points), dtype=np.intp)
+from spatial_index import SphereIndexBuilder
 
 
-def test_add_points():
-    print("Running tests with class " + core.SphereIndex.__name__)
-    t = core.SphereIndex()
-    t._add_points(points, ids)
-    min_corner = [-1, -1, -1]
-    max_corner = [1, 1, 1]
-    idx = _box_query_id(t, min_corner, max_corner)
-    expected_result = np.array([0, 1, 2, 6], dtype=np.uintp)
-
-    assert sorted(idx) == sorted(expected_result)
-
-
-def test_init_points():
-    p = core.SphereIndex(points, radii=None)  # or p = core.SphereIndex(points, None)
-    idx = p._find_nearest([5, 0, 0], 3)
-    expected_result = np.array([0, 2, 6], dtype=np.uintp)
-    assert np.all(idx == expected_result), (idx, expected_result)
-
-
-def test_init_points_with_ids():
-    p = core.SphereIndex(points, radii=None, py_ids=ids)
-    # or p = core.SphereIndex(points, None, ids)
-    idx = p._find_nearest([5, 0, 0], 3)
-    expected_result = np.array([0, 2, 6], dtype=np.uintp)
-    assert np.all(idx == expected_result), (idx, expected_result)
-
-
-def test_print_rtree():
-    points = np.array(
+@pytest.fixture
+def points():
+    return np.array(
         [
-            [0.00001, 1.0, 0],
+            [1e-5, 1.0, 0],
             [-0.5, -0.5456, 0],
             [0.5, -0.5, 0],
-        ], dtype=np.float32)
-    p = core.SphereIndex(points[0:3], radii=None)
+            [-2.1, 0.0, 0],
+            [0.0, 2.1, 0],
+            [0.0, 0.0, 2.1],
+            [1., 1., 1.],
+        ],
+        dtype=np.float32,
+    )
+
+
+@pytest.fixture
+def ids(points):
+    return np.arange(len(points), dtype=np.intp)
+
+
+@pytest.fixture
+def zero_radii(points):
+    return np.zeros(points.shape[0], dtype=np.float32)
+
+
+def test_init_points_with_ids(points, zero_radii, ids):
+    index = SphereIndexBuilder.from_numpy(points, zero_radii, ids)
+
+    box = [-1, -1, -1], [1, 1, 1]
+    actual_result = index.box_query(*box, fields="id")
+    expected_result = np.array([0, 1, 2, 6], dtype=np.uintp)
+
+    assert sorted(actual_result) == sorted(expected_result)
+
+
+def test_print_rtree(points, zero_radii, ids):
+    p = core.SphereIndex(points[:3, :], zero_radii[:3], ids[:3])
     str_expect = (
         'IndexTree([\n'
         '  IShape(id=0, Sphere(centroid=[1e-05 1 0], radius=0))\n'